+++ lib/pleroma/upload/filter/exiftool.ex
@@ -1,0 +1,17 @@
+# Pleroma: A lightweight social networking server
+# Copyright © 2017-2020 Pleroma Authors <https://pleroma.social/>
+# SPDX-License-Identifier: AGPL-3.0-only
+
+defmodule Pleroma.Upload.Filter.Exiftool do
+  @behaviour Pleroma.Upload.Filter
+
+  @type conversion :: action :: String.t() | {action :: String.t(), opts :: String.t()}
+  @type conversions :: conversion() | [conversion()]
+
+  def filter(%Pleroma.Upload{tempfile: file, content_type: "image" <> _}) do
+    System.cmd("exiftool", ["-overwrite_original", "-gps:all=", file], parallelism: true)
+    :ok
+  end
+
+  def filter(_), do: :ok
+end